Create README.md
He is an interdisciplinary Industrial Engineer that have carried out a process of reorientation from the industrial process improvement sector, towards the digital/technological sector. For this, He is studying a master's degree in Data Science, also last year a postgraduate degree in Digital Transformation Management and previously a master's degree in Internet of Things.

He wants to find a nice challenge in a company where he can continue developing his professional career in the digital/technological field, either as a senior or starting as a junior, depending on the position, as recognizes that he is new to the data/technology field.

He lived in Birmingham (UK) for almost three years working as a Process Engineer in a factory, where coordinated international and multidisciplinary teams.

He thinks he is considered a great professional, proactive and unafraid to make decisions.

His values are honesty, companionship, commitment, and rigor.